cleanup helpers and resolve css and js loading issue
