+++ Gemfile
@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 ruby "2.5.5"
-
+#another try
 gem 'rails', '5.2.3'
 gem "dotenv-rails"
 gem "dotenv-deployment"
@@ -17,9 +17,9 @@ group :development,:test, :integration do
   gem "spreewald", :git => 'https://github.com/makandra/spreewald.git'
 end
 
-#group :production,:staging do
-#  gem 'mysql'
-#end
+group :production,:staging do
+  gem 'mysql2'
+end
 
 gem 'savon', '~> 2.11.1'
 gem 'parslet'
@@ -30,7 +30,7 @@ gem 'blacklight_range_limit', '~> 7.0'
 gem 'blacklight_unapi', :git => 'https://github.com/cul-it/blacklight-unapi', :branch => 'BL7-upgrade'
 gem 'kaminari', '>= 0.15'
 
-gem 'blacklight-hierarchy', :path => '/usr/local/src/bc-hierarchy'
+gem 'blacklight-hierarchy', :git => 'https://github.com/cul-it/blacklight-cornell-hierarchy', :branch => 'master'
 gem 'htmlentities'
 gem 'json'
 gem 'httpclient'
@@ -100,10 +100,12 @@ gem 'omniauth-saml'
 gem 'omniauth-google-oauth2', '~> 0.4.1'
 # Use Omniauth Facebook plugin
 gem 'omniauth-facebook', '~> 4.0'
-
+gem 'xmlrpc'
 gem 'bootstrap'
 gem 'font-awesome-rails'
 gem 'blacklight_cornell_requests',:git =>'https://github.com/cul-it/blacklight-cornell-requests', :branch => 'BL7-upgrade'
+#gem 'blacklight_cornell_requests', :path => '/Users/matt/code/d&a/blacklight-cornell-requests'
+#gem 'my_account', :git => 'https://github.com/cul-it/cul-my-account', :branch => 'dev'
 gem 'borrow_direct', :git => 'https://github.com/jrochkind/borrow_direct'
 
 gem 'bento_search'
@@ -117,6 +119,3 @@ gem 'csl-styles', :git => 'https://github.com/cul-it/csl-styles', :branch => 'ma
 gem 'citeproc-ruby'
 gem 'unicode_utils'
 gem 'google-analytics-rails', '1.1.1'
-
-# for Rails 5
-gem 'xmlrpc'
