+++ app/assets/javascripts/application.js
@@ -10,25 +10,24 @@
 // WARNING: THE FIRST BLANK LINE MARKS THE END OF WHAT'S TO BE PROCESSED, ANY BLANK LINE SHOULD
 // GO AFTER THE REQUIRES BELOW.
 //
+// require bootstrap-sprockets can likely delete this
+
+// Required by Blacklight
 //= require jquery
-//= require bootstrap-sprockets
-
-
-
-
-
+//= require popper
+//= require bootstrap
+//= require blacklight/blacklight
 
 
 // For blacklight_range_limit built-in JS, if you don't want it you don't need
 // this:
 //= require 'blacklight_range_limit'
-//  require 'range_limit_slider.js'
+
+
 //= require 'blacklight/hierarchy/hierarchy'
 //= require jquery_ujs
 //= require jquery-ui/autocomplete
 //
-// Required by Blacklight
-//= require blacklight/blacklight
 //= require_tree .
 //
 //= require blacklight_cornell_requests
@@ -43,3 +42,4 @@
 //= require jquery.waypointz.js
 //= require zz_inview.js
 //= require tooltip.js
+
