+++ app/assets/stylesheets/application.css.scss
@@ -18,12 +18,8 @@
  *
  * Used by blacklight_range_limit
  *= require  'blacklight_range_limit'
+ *
  *= require jquery-ui
- *
-
- *
- * Used by blacklight_range_limit
- *= require  'blacklight_range_limit'
  *
 */
 // Cornell catalog variables & overrides
