+++ app/views/catalog/show.html.erb
@@ -1,8 +1,6 @@
 <% if back_to_catalog_needed %>
   <% session[:search][:controller] == 'advanced_search' ? link_label = 'Back to results' : '' %>
   <% link_back = link_back_to_catalog({:label => link_label}) %>
-<% Rails.logger.info("LINK BACK = " + link_back.inspect)%>
-<% Rails.logger.info("LINK BACK = " + link_back.class.inspect)%>
 <% if link_back and link_back.class.to_s != "ActiveSupport::SafeBuffer"%>
 <%# if link_back %>
     <% if !link_back[:url].include? "databases" and link_back[:url] != "/"  %>
